Unsing type off_t instead of int for InputStream#available().
git-svn-id: https://svn.code.sf.net/p/stemkit/code@89 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ include/sk/io/AbstractInputStream.h
@@ -29,7 +29,7 @@ namespace sk {
         std::vector<char> read(int number);
         std::vector<char>& read(std::vector<char>& buffer, int number);
         int skip(int number);
-        int available() const;
+        off_t available() const;
         bool markSupported() const;
         void mark(int readlimit);
         void reset();
